Rename type_parameter to type_param.
